added logging, code refactor
